Conflict in GraphicMain
